SphericalBesselFunctions.jl

SphericalBesselFunctions.bessel_fraction

SphericalBesselFunctions.bessels!

SphericalBesselFunctions.lentz_thompson

SphericalBesselFunctions.powneg1

SphericalBesselFunctions.reflect!

SphericalBesselFunctions.bessels! — Method

bessels!(j, j′, y, y′, x)

Compute all spherical Bessel (regular: $j_n$) and Neumann (irregular: $y_n$) functions and their derivatives at x and store the results in the vectors j, j′, y, y′. If only the Bessel functions or the Neumann functions are of interest, the other pair of arrays can be substituted by nothing. However, it is not possible to compute only the functions but not the derivatives, since they are generated using the following recurrence relations:

\[\begin{aligned} g_{n-1} &= \frac{n+1}{x} g_n + g_n'\\ g'_{n-1} &= \frac{n-1}{x} g_{n-1} - g_n. \end{aligned}\]

These recurrence relations are employed in a downward fashion for the Bessel functions and an upward fashion for the Neumann functions.

It is assumed that all passed arrays are of the same lengths (not checked).

SphericalBesselFunctions.bessel_fraction — Method

bessel_fraction(x, n)

Evaluated the continued fraction for the spherical Bessel function

\[\frac{j'_n(x)}{j_n(x)} = \frac{n}{x} - \frac{1}{T_{n+1}(x)-}\frac{1}{T_{n+2}(x)-}...\frac{1}{T_k-...},\]

where

\[T_k(x) = \frac{2k+1}{x}.\]

SphericalBesselFunctions.lentz_thompson — Method

lentz_thompson(b₀, a, b)

Lentz–Thompson algorithm for the forward evaluation of continued fractions:

\[f_n = b_0 + \frac{a_1}{b_1+}\frac{a_2}{b_2+}...\frac{a_{n-1}}{b_{n-1}+}\frac{a_n}{b_n}\]

As described in

Barnett, A. R. (1996). The Calculation of Spherical Bessel and Coulomb Functions. In (Eds.), Computational Atomic Physics (pp. 181–202). Springer Berlin Heidelberg.

SphericalBesselFunctions.powneg1 — Method

powneg1(m)

Returns an integer power of negative unity, i.e. $(-)^n$.

SphericalBesselFunctions.reflect! — Function

reflect!(g, g′, offset=0)

Affect the reflection symmetries

\[j_n(-z) = (-)^n j_n(z), \qquad y_n(-z) = (-)^{n+1}y_n(z),\]

where the $+1$ can be added using offset.